+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/auth/social/data/KakaoTokenInfo.kt
@@ -1,3 +1,3 @@
-package siksha.wafflestudio.core.domain.auth.social
+package siksha.wafflestudio.core.domain.auth.social.data
 
 data class KakaoTokenInfo(val id: Long, val expires_in: Long, val app_id: Long)
