+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/common/config/OAuthConfig.kt
@@ -2,7 +2,7 @@ package siksha.wafflestudio.core.domain.common.config
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration
-import siksha.wafflestudio.core.domain.auth.social.GoogleOauthProperties
+import siksha.wafflestudio.core.domain.auth.social.data.GoogleOauthProperties
 
 @EnableConfigurationProperties(GoogleOauthProperties::class)
 @Configuration
